Crate atomic_arena

atomic_arena provides a generational Arena that you can reserve a Key for ahead of time using a Controller. Controllers are backed by atomics, so they can be cloned and used across threads and still have consistent state.

This is useful when you want to insert an item into an Arena on a different thread, but you want to have a valid Key for that item immediately on the current thread.

Modules

error

Error types.

iter

Arena iterators.

Structs

Arena

A container of items that can be accessed via a Key.

Controller

Manages Key reservations for an Arena.

Key

A unique identifier for an item in an Arena.